Add tagging functionality
This adds the ability to push tags to the upstream remote
by running a gradle task. This will calculate the version,
tag the version locally and then push up the version tag.

You can configure this tagging functionality by using the
tag configuration block in the build file. Currently you
can customise the prefix used for the tag, which defaults
to v.