Add authentication methods
This adds authentication for SSH and HTTPS.

SSH should work anyway, but currently JGit only supports RSA keys
for known hosts, and will fail for ecdsa keys. The only real
workaround for now is to disable the strict host checking for
ecdsa keys if you can't use RSA.

HTTPS is a bit easier. You can use username, password, and token.
Username and password work as expected and pass along the
credentials with the request. Token probably isn't the best
implementation it could be- it's been built with GitHub in mind,
so it will pass the token in as the username for the request. This
might not work for other platforms though. I don't have that use
case right now though so we can just wait for an issue to be raised
with that specific requirement and implement it then.