Add version pattern capability
This adds a version object that contains logic for printing the version
according to the supplied pattern.

Patterns are simple at the minute to support simple use cases. They
can probably be made more complex in the future if required.

This still needs to be plumbed in to allow consumers to properly utilise
it in their projects.