Update versioner to handle new versioning logic
This updates the versioner to return the new Version object instead of
a string. This removes the need to have any pattern matching logic in
the versioner like there was before.

This will cause the other parts of the system to behave weirdly until
they're updated to work with this new change.

I've also added the branch and commit hash to the version object so we
can use them in the patterns.