Update configuration for consumer provided patterns
This wires in the pattern configuration item to all of the necessary
tasks, to allow them to calculate and print the version according to the
specified pattern.

The pattern defaults to %M.%m.%p(.%c) which matches what versioner
would have outputted before the pattern work took place. This is to
allow the change to be backwards compatible, though I would prefer to
eventually change the default pattern to be something a bit more semver
friendly. But not yet.