Change build references to commit
I think this makes it more obvious- 'build number' was a bit more
opaque than 'commit number'. This should make it more obvious what we're
referring to at a glance.

It also freed up using %b for branch instead of having to use %B and %b
if we were still using build.