Add full hash functionality
This should make it easier to link a version back to an exact commit
in URLs and things like that. This maintains the short hash pattern,
though it is no longer guaranteed to be a unique short hash, only that
it is the first seven characters of the current hash.