Add actions in addition to closures for nested configuration
I tried replacing the closures with actions as described in the gradle
documentation- this worked for kotlin scripts but the methods were not
able to be called from groovy scripts.

Having both a closure and an action seems to work the best, groovy
scripts will call into the closure methods and kotlin scripts call into
the action methods.