Add extra test types
This adds the concept of functional and integration tests that we can run outside of
the test source set. This follows the Gradle plugin testing recommendations to
perform these types of tests.

Also updates the version of kotlintest, but still uses the deprecated methods. All
of the test classes will be touched in this feature branch so we can update the
warnings at that point.

Removes AssertJ as we're exclusively using kotlintest assertions now, or will be by the
time this is over.