Add integration tests for GitTagger
This adds some tests for the git tagging functionality as integration tests to
GitTagger.

Took this opportunity to tidy up GitTagger a ~bit~ lot too, and make it more testable and
less reliant on Gradle project specific objects.

This removes the reliance on the Gradle specific configuration objects, and replaced
them with some "core" configuration with a flatter hierarchy that is constructed using
the properties of the Gradle configuration objects directly.

This required tidying up some other places including the Plugin class, and the tagging
task. These will get a more extensive tidy as I add more tests.