Add tests for Versioner
This adds some tests for the versioning functionality as integration tests to
Versioner.

Like before I took some time to refactor the Versioner to be less reliant on
Gradle objects to make it easier to test. I'm not 100% sure I've done the
right thing here so may end up changing it again- I've been sitting on this
commit since 2019 so I've probably learnt more about plugins in that time.