Update configuration to actually work
This changes the configuration around versioning to actually pick up the
configuration from the extension and apply it to the versioner.

This also changes the match pattern to not be a regular expression any more,
which changes the functionality of the plugin but this wasn't expressly
advertised before so might be ok.

Otherwise everything works as before.