Only version once
Before we were versioning every time we applied the plugin, and then every
time we ran a task.

Now we only version on plugin application and use that version everywhere else
instead of calculating it again.

This means a lot of refactoring of the existing structure and tests, which
has been merged in with the configuration refactoring of the GitTagger because
I forgot to separate them properly. My bad.

I might do a proper refactor after all this is finished at the end to put all
of my ducks in a consistent row.