Update to latest jgit (5.9.0)
This has some breaking changes in it- namely that it seems to sign by default
now. Or maybe that's just my local git configuration taking effect. Anyway,
I've had to explicitly not sign in the tests to get it to work.

Also no longer includes Jsch. I've included it explicitly as we need it- but
we should have had this before anyway as we shouldn't have relied on the
transitive dependeny to be there. Because then it wasn't.